feat(kikan): lift login/logout/me to /api/platform/v1/auth (#685)
Promotes the three simplest auth handlers (login, logout, me) from
crates/mokumo-shop/src/auth_handlers/ into the kikan platform layer:

- New axum adapters at crates/kikan/src/platform/v1/auth/{login,logout,me}.rs.
  Generic over K = Graft::ProfileKind via the ProfileKindBounds trait alias.
  Auth pool sourced from PlatformState::auth_profile_kind_dir (string
  snapshot of Graft::auth_profile_kind() taken at boot) — no vertical
  literals leak into kikan.
- New auth_router<S, K>(control_plane) builder mounting the canonical URLs
  /api/platform/v1/auth/{login,logout,me}. mokumo-shop merges this into its
  data-plane routes.
- Legacy /api/auth/{login,logout,me} alias preserved by mokumo-shop's
  routes.rs by routing the legacy paths to the same kikan handlers via
  turbofish ::<SetupMode>. Demo-auto-login middleware (require_auth_with_
  demo_auto_login) continues to wrap the legacy /api/auth/me path on the
  shop side per Decision D4 — the canonical /api/platform/v1/auth/me does
  NOT pass through it (admin UI sessions are explicit).
- mokumo-shop/src/auth_handlers/mod.rs reduced to legacy-alias scaffolding
  (recover/reset/setup/regenerate-codes still shop-side; lifted in later
  commits). Removed: login/logout/me handlers, user_to_response helper,
  is_still_locked, handle_failed_login, LOGIN_LOCKOUT_THRESHOLD/SECS, and
  the auth_router/auth_me_router builders.

Hurl smokes added at tests/api/platform/auth/{login-success,login-bad-
credentials,login-rate-limit,me,me-unauthenticated,logout}.hurl exercising
the canonical URLs directly. login-rate-limit uses a distinct fixture email
(ratelimit-platform-smoke@smoke.invalid) so the in-memory limiter stays
isolated from the legacy alias smoke (ratelimit-smoke-test@smoke.invalid),
since both URLs share the same handler and limiter.

Invariants verified:
- I1 vocabulary purity: scripts/check-i1-vocabulary-purity.sh passes
- I1 domain purity: scripts/check-i1-domain-purity.sh passes
- I4 DAG: cargo tree -p kikan shows no mokumo-shop edge
- All existing legacy-alias BDD scenarios continue to pass via the alias
  mount; demo_auth scenarios confirm middleware wrapping still functions

Pre-existing baseline BDD failures (26 of 206 scenarios — DefensiveEmpty
ShopName boot errors, restore tests, shop_logo tests, recovery PIN
re-attempt) are unchanged by this commit; verified by stash + re-run.

Refs #685.
