+++ crates/mokumo-shop/src/routes.rs
@@ -14,7 +14,7 @@ use axum::routing::{get, post};
 use axum::{Json, Router};
 
 use kikan::AppError;
-use kikan_types::HealthResponse;
+use kikan_types::{HealthResponse, SetupMode};
 
 use crate::state::SharedMokumoState;
 
@@ -46,8 +46,18 @@ pub fn data_plane_routes(state: &SharedState) -> Router<SharedState> {
     );
 
     // ── Protected auth sub-router ───────────────────────────────────
+    //
+    // Legacy `/api/auth/me` alias is mounted here behind
+    // `require_auth_with_demo_auto_login` so demo-mode sessions auto-issue
+    // before the auth check (Mokumo-specific policy — see ADR
+    // `adr-platform-auth-handler-placement`). The canonical
+    // `/api/platform/v1/auth/me` mount is kikan-side and does NOT pass
+    // through demo-auto-login: admin-UI sessions are always explicit.
     let protected_auth_routes = Router::new()
-        .nest("/api/auth", crate::auth_handlers::auth_me_router())
+        .route(
+            "/api/auth/me",
+            get(kikan::platform::v1::auth::me::me::<SetupMode>),
+        )
         .route(
             "/api/account/recovery-codes/regenerate",
             post(crate::auth_handlers::regenerate_recovery_codes),
@@ -108,12 +118,35 @@ pub fn data_plane_routes(state: &SharedState) -> Router<SharedState> {
         )
         .layer(axum::extract::DefaultBodyLimit::max(500 * 1024 * 1024));
 
+    // ── Public auth: legacy /api/auth alias + recover ───────────────
+    //
+    // Legacy `/api/auth/{login,logout}` route to the same kikan handlers
+    // as the canonical `/api/platform/v1/auth/{login,logout}` mount; the
+    // shop SPA continues to call the alias through M0. `/api/auth/recover`
+    // is shop-vertical (it depends on Mokumo's recovery-code wire shape)
+    // and stays here.
+    let public_auth_router = Router::new()
+        .route(
+            "/login",
+            post(kikan::platform::v1::auth::login::login::<SetupMode>),
+        )
+        .route(
+            "/logout",
+            post(kikan::platform::v1::auth::logout::logout::<SetupMode>),
+        )
+        .route("/recover", post(crate::auth_handlers::recover::recover))
+        .with_state(control_plane_state.clone());
+
     // ── Public routes ───────────────────────────────────────────────
     let mut router = Router::new()
         .route("/api/health", get(health))
         .merge(kikan::data_plane::kikan_version::kikan_version_router::<
             SharedState,
         >(state.platform_state()))
+        .merge(kikan::platform::v1::auth::auth_router::<
+            SharedState,
+            SetupMode,
+        >(control_plane_state.clone()))
         .route("/api/server-info", get(crate::server_info::handler))
         .route("/api/setup-status", get(setup_status))
         .route(
@@ -126,9 +159,7 @@ pub fn data_plane_routes(state: &SharedState) -> Router<SharedState> {
         )
         .nest(
             "/api/auth",
-            crate::auth_handlers::auth_router()
-                .with_state(control_plane_state.clone())
-                .merge(crate::auth_handlers::reset_router()),
+            public_auth_router.merge(crate::auth_handlers::reset_router()),
         )
         .nest(
             "/api/setup",
